infra-awscdk-app2-stacks/requirements.txt
--extra-index-url https://${FURY_AUTH}:@pypi.fury.io/<your-company>/
infra-awscdk-devops-packages==0.1.32
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
python-dotenv>=1.0.1
pyyaml==6.0.1
aws-cdk.lambda-layer-kubectl-v30==2.0.0